requirements.txt
azure-core==1.20.1
azure-storage-blob==12.9.0
